Fix review lookup url
